# swagger.json
{
  "components": {
    "schemas": {
...
      "RefundStatus": {
        "enum": [
          "REQUESTED",
          "APPROVED",
          "DENIED",
          "PROCESSED",
          "COMPLETED",
          "CANCELLED"
        ],
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Comprehensive API for bus payment system with offers, rewards, bookings, and refunds",
    "title": "Bus Payments & Rewards API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/refunds": {
      "get": {
        "description": "Get all refund requests with optional filtering",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "default": 10,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "$ref": "#/components/schemas/RefundStatus"
            }
          },
          {
            "in": "query",
            "name": "customer_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "refund_type",
            "schema": {
              "enum": [
                "booking_cancellation",
                "dispute_resolution",
                "goodwill"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of refunds with pagination"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "List refunds",
        "tags": [
          "Refunds"
        ]
      }
    },
...
    "/api/refunds/{refund_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "refund_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Refund details"
          },
          "404": {
            "description": "Refund not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get refund details",
        "tags": [
          "Refunds"
        ]
      }
    },
    "/api/refunds/{refund_id}/approve": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "refund_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Refund approved successfully"
          },
          "400": {
            "description": "Refund cannot be approved"
          },
          "404": {
            "description": "Refund not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Approve refund",
        "tags": [
          "Refunds"
        ]
      }
    },
    "/api/refunds/{refund_id}/deny": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "refund_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "denial_reason": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Refund denied successfully"
          },
          "400": {
            "description": "Refund cannot be denied"
          },
          "404": {
            "description": "Refund not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Deny refund",
        "tags": [
          "Refunds"
        ]
      }
    },
    "/api/rewards": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "customer_id": {
                    "type": "integer"
                  },
                  "description": {
                    "type": "string"
                  },
                  "dollar_value": {
                    "type": "number"
                  },
                  "expiry_date": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "offer_id": {
                    "type": "integer"
                  },
                  "points_earned": {
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "customer_id",
                  "points_earned",
                  "description"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Reward created successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "404": {
            "description": "Customer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create manual reward",
        "tags": [
          "Rewards"
        ]
      }
    },
...
    "/api/rewards/{reward_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "reward_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reward details"
          },
          "404": {
            "description": "Reward not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get specific reward",
        "tags": [
          "Rewards"
        ]
      }
    },
    "/api/rewards/{reward_id}/redeem": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "reward_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "points": {
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Reward points redeemed successfully"
          },
          "400": {
            "description": "Invalid redemption request"
          },
          "404": {
            "description": "Reward not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Redeem specific reward",
        "tags": [
          "Rewards"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://192.168.86.189:5001"
    }
  ]
}